Fixed classes effected by the package move
