+++ src/main/groovy/liquibase/serializer/ext/GroovyChangeLogSerializer.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -33,6 +33,7 @@ import java.sql.Timestamp
  *  found by Liquibase at runtime.
  *
  * @author Tim Berglund
+ * @author Steven C. Saliman
  */
 class GroovyChangeLogSerializer
 				implements ChangeLogSerializer {
