+++ src/main/groovy/liquibase/util/PatchedObjectUtil.groovy
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
 package liquibase.util;
 
 import liquibase.exception.UnexpectedLiquibaseException;
@@ -20,6 +35,9 @@ import java.util.Map;
  * silently ignoring them.  See
  * https://liquibase.jira.com/browse/CORE-1968?focusedCommentId=24201#comment-24201
  * for mor information.
+ *
+ * @author Nathan Voxland
+ * @author Steven C. Saliman
  */
 class PatchedObjectUtil {
