+++ src/main/groovy/org/liquibase/groovy/delegate/ChangeSetDelegate.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package net.saliman.liquibase.delegate
+package org.liquibase.groovy.delegate
 
 import liquibase.change.AddColumnConfig
 import liquibase.change.ColumnConfig
@@ -39,6 +39,8 @@ import liquibase.util.PatchedObjectUtil
  * table name is required.  Similarly, it doesn't make sense to have an
  * addColumn change without at least one column, so we don't deal well with the
  * addColumn change without a closure.
+ *
+ * @author Steven C. Saliman
  */
 class ChangeSetDelegate {
 	def changeSet
