+++ src/main/groovy/org/liquibase/groovy/delegate/DatabaseChangeLogDelegate.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package net.saliman.liquibase.delegate
+package org.liquibase.groovy.delegate
 
 import liquibase.ContextExpression
 import liquibase.Labels
@@ -26,7 +26,6 @@ import liquibase.parser.ChangeLogParserFactory
  * This class is the delegate for the {@code databaseChangeLog} element.  It
  * is the starting point for parsing the Groovy DSL.
  *
- * @author Tim Berglund
  * @author Steven C. Saliman
  */
 class DatabaseChangeLogDelegate {
