+++ src/main/groovy/org/liquibase/groovy/delegate/KeyValueDelegate.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package net.saliman.liquibase.delegate
+package org.liquibase.groovy.delegate
 
 import liquibase.exception.ChangeLogParseException
 
@@ -31,7 +31,6 @@ import liquibase.exception.ChangeLogParseException
  * The map created by this delegate will not do database changeLog property
  * substitution, that will be up to the caller.
  *
- * @author Tim Berglund
  * @author Steven C. Saliman
  */
 class KeyValueDelegate {
