+++ src/main/groovy/org/liquibase/groovy/liquibase.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package net.saliman.liquibase
+package org.liquibase.groovy
 
 import liquibase.parser.ext.GroovyLiquibaseChangeLogParser
 import liquibase.parser.*
