+++ src/test/groovy/org/liquibase/groovy/delegate/ArchitecturalRefactoringTests.groovy
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package net.saliman.liquibase.delegate
+package org.liquibase.groovy.delegate
 
 import liquibase.exception.ChangeLogParseException
 import org.junit.Test
@@ -57,7 +57,6 @@ import liquibase.change.core.DropIndexChange
  * least one column.  If a user tries it, they will get errors from Liquibase
  * itself.
  *
- * @author Tim Berglund
  * @author Steven C. Saliman
  */
 class ArchitecturalRefactoringTests extends ChangeSetTests {
