+++ src/test/groovy/org/liquibase/groovy/delegate/ChangeSetPreconditionTests.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,9 +14,9 @@
  *  limitations under the License.
  */
 
-package net.saliman.liquibase.delegate
+package org.liquibase.groovy.delegate
 
-import net.saliman.liquibase.delegate.ChangeSetTests
+import org.liquibase.groovy.delegate.ChangeSetTests
 import org.junit.Test
 import static org.junit.Assert.*
