+++ src/test/groovy/org/liquibase/groovy/delegate/ColumnDelegateTests.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package net.saliman.liquibase.delegate
+package org.liquibase.groovy.delegate
 
 import liquibase.change.AddColumnConfig
 import liquibase.change.ColumnConfig
@@ -35,7 +35,6 @@ import java.text.SimpleDateFormat
  * that we can pass things to Liquibase correctly. We check all attributes that
  * are known at this time - note that several are undocumented.
  *
- * @author Tim Berglund
  * @author Steven C. Saliman
  */
 class ColumnDelegateTests {
