+++ src/test/groovy/org/liquibase/groovy/delegate/DelegateExtensionTests.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,15 +14,15 @@
  *  limitations under the License.
  */
 
-package net.saliman.liquibase.delegate
+package org.liquibase.groovy.delegate
 
 import static org.junit.Assert.*
 import liquibase.change.core.RawSQLChange
 
 import org.junit.Test
 
-import net.saliman.liquibase.change.CustomProgrammaticChangeWrapper
-import net.saliman.liquibase.custom.MyCustomSqlChange
+import org.liquibase.groovy.change.CustomProgrammaticChangeWrapper
+import org.liquibase.groovy.custom.MyCustomSqlChange
 
 
 /**
