+++ src/test/groovy/org/liquibase/groovy/delegate/PreconditionDelegateTests.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package net.saliman.liquibase.delegate
+package org.liquibase.groovy.delegate
 
 import liquibase.exception.ChangeLogParseException
 import liquibase.precondition.core.NotPrecondition;
@@ -47,7 +47,6 @@ import liquibase.precondition.CustomPreconditionWrapper
  * since we just pass through to Liquibase, we're not too concerned with
  * validating attrubutes.
  *
- * @author Tim Berglund
  * @author Steven C. Saliman
  */
 class PreconditionDelegateTests {
