+++ src/test/groovy/org/liquibase/groovy/delegate/StructuralRefactoringTests.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package net.saliman.liquibase.delegate
+package org.liquibase.groovy.delegate
 
 import liquibase.change.core.AddColumnChange
 import liquibase.change.core.ModifyDataTypeChange
@@ -69,7 +69,6 @@ import liquibase.change.core.CreateProcedureChange
  * least one column.  If a user tries it, they will get errors from Liquibase
  * itself.
  *
- * @author Tim Berglund
  * @author Steven C. Saliman
  */
 class StructuralRefactoringTests extends ChangeSetTests {
