+++ src/test/groovy/org/liquibase/groovy/serialize/ArchitecturalRefactoringSerializerTests.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package net.saliman.liquibase.serialize
+package org.liquibase.groovy.serialize
 
 import org.junit.Test
 import static org.junit.Assert.*
@@ -23,9 +23,7 @@ import liquibase.change.core.CreateIndexChange
 import liquibase.change.core.DropIndexChange
 
 
-class ArchitecturalRefactoringSerializerTests
-  extends SerializerTests
-{
+class ArchitecturalRefactoringSerializerTests extends SerializerTests {
 
   @Test
   void createIndexWithMultipleColumns() {
