+++ src/test/groovy/org/liquibase/groovy/serialize/SerializerTests.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package net.saliman.liquibase.serialize
+package org.liquibase.groovy.serialize
 
 import liquibase.resource.FileSystemResourceAccessor
 import liquibase.serializer.ChangeLogSerializerFactory
@@ -28,7 +28,7 @@ import java.sql.Timestamp
 /**
  * A base class providing support for test classes targeting serializations.
  *
- * @author Tim Berglund
+ * @author Steven C. Saliman
  */
 class SerializerTests {
   def sdf = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.s")
