+++ src/test/groovy/org/liquibase/groovy/serialize/SqlVisitorSerializerTests.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package net.saliman.liquibase.serialize
+package org.liquibase.groovy.serialize
 
 import org.junit.Test
 import static org.junit.Assert.*
@@ -23,11 +23,9 @@ import liquibase.sql.visitor.ReplaceSqlVisitor
 /**
  * <p></p>
  * 
- * @author Tim Berglund
+ * @author Steven C. Saliman
  */
-class SqlVisitorSerializerTests
-  extends SerializerTests
-{
+class SqlVisitorSerializerTests extends SerializerTests {
 
   @Test
   void testReplaceSqlVisitorSerialize() {
