+++ src/test/groovy/org/liquibase/precondition/MonkeyWinPrecondition.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2014 Tim Berglund and Steven C. Saliman
+ * Copyright 2011-2015 Tim Berglund and Steven C. Saliman
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -20,9 +20,7 @@ import liquibase.precondition.CustomPrecondition
 import liquibase.database.Database
 
 
-class MonkeyWinPrecondition
-  implements CustomPrecondition
-{
+class MonkeyWinPrecondition implements CustomPrecondition {
 
   void check(Database database) {
     // Do nothing, and the precondition passes
